[refactor] Move formats to lib folder
