+++ lib/align.js
