+++ lib/browser.js
