+++ lib/cli.js
