+++ lib/colorize.js
