+++ lib/combine.js
