+++ lib/format.js
