+++ lib/label.js
