+++ lib/logstash.js
