+++ lib/metadata.js
