+++ lib/ms.js
