+++ lib/printf.js
