+++ lib/splat.js
