+++ lib/timestamp.js
