+++ lib/uncolorize.js
