+++ test/simple.test.js
@@ -2,7 +2,7 @@
 'use strict';
 
 const assume = require('assume');
-const simple = require('../simple');
+const simple = require('../lib/simple');
 const helpers = require('./helpers');
 const { MESSAGE } = require('triple-beam');
