+++ test/uncolorize.test.js
@@ -3,11 +3,11 @@
 
 const assume = require('assume');
 const colors = require('colors');
-const colorize = require('../colorize');
-const combine = require('../combine');
-const format = require('../format');
-const simple = require('../simple');
-const uncolorize = require('../uncolorize');
+const colorize = require('../lib/colorize');
+const combine = require('../lib/combine');
+const format = require('../lib/format');
+const simple = require('../lib/simple');
+const uncolorize = require('../lib/uncolorize');
 const helpers = require('./helpers');
 const { MESSAGE } = require('triple-beam');
 const COLORED = Symbol.for('colored');
